packages/agents/pyproject.toml
[project]
name = "youtube-ai-agents"
version = "0.1.0"
description = "YouTube AI Agent Agency - LangGraph 기반 멀티 채널 자동화 시스템"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "langgraph-supervisor>=0.0.31",
]

[project.optional-dependencies]
media = [
    "elevenlabs>=1.0.0",
    "moviepy>=2.0.0",
    "openai-whisper>=20240930",
]
publisher = [
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
]
search = [
    "tavily-python>=0.5.0",
    "beautifulsoup4>=4.12.0",
]
dashboard = []
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]
all = [
    "youtube-ai-agents[media,publisher,search,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["N802"]
